requirements.txt
flask
flask-cors
ultralytics==8.3.241
torch==2.0.1+cpu
torchvision==0.15.2+cpu
numpy==1.26.4
pillow
opencv-python==4.9.0.80
